+++ src/lib/lforms/lhc-form-utils.js
@@ -70,17 +70,15 @@ const FormUtils = {
         eleLhcForm.addEventListener('onFormReady', function(e){
           resolve()
         });        
+        eleLhcForm.addEventListener('onError', function(e){
+          reject(e.detail)
+        });        
       }
       catch(e) {
         reject(e)
       }
     }).catch(function(e) {
-      var errMsg = (e instanceof Error) ? e.message : e.toString();
-      const divMessage = document.createElement("div");    
-      divMessage.innerHTML = '<div id=lformsErrors style="color: red"><b>Unable to '+
-      'display the form, due to the following error:</b><p id=lformsErrorContent>' + errMsg + '</p>';
-      formContainer.appendChild(divMessage)
-      throw e;
+      throw e; 
     });
 
     return rtnPromise;